requirements.txt
fastapi
uvicorn
asyncpg
pydantic
pydantic-settings
python-dotenv
python-multipart
google-genai
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
twilio
aiokafka
pytest
pytest-asyncio
httpx
openai
